+++ src/test/java/janggi/piece/pieces/RunningPiecesTest.java
@@ -1,10 +1,16 @@
-package janggi.piece;
+package janggi.piece.pieces;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import janggi.game.Team;
+import janggi.game.team.Team;
+import janggi.piece.Byeong;
+import janggi.piece.Cha;
+import janggi.piece.Gung;
+import janggi.piece.Ma;
+import janggi.piece.Movable;
+import janggi.piece.Po;
 import janggi.point.Point;
 import java.util.HashMap;
 import java.util.Map;
