+++ src/test/java/janggi/point/PalacePointsTest.java
@@ -3,7 +3,7 @@ package janggi.point;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import janggi.game.Team;
+import janggi.game.team.Team;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
